Put new prompt templates under MIT
